requirements.txt
pillow
tkinter
sqlalchemy
